Module Closure_conversion_aux

Environments and auxiliary structures used during closure conversion.

module Env : sig ... end

Used to remember which Variable.t values correspond to which Ident.t values during closure conversion, and similarly for static exception identifiers.

module Function_decls : sig ... end

Used to represent information about a set of function declarations during closure conversion. (The only case in which such a set may contain more than one declaration is when processing "let rec".)