Module Cmm_invariants

run ppf fundecl analyses the given function, and returns whether any errors were encountered (with corresponding error messages printed on the given formatter).

val run : Format.formatter -> Cmm.fundecl -> bool