Module Remove_free_vars_equal_to_args

Replace free variables in closures known to be equal to specialised arguments of such closures with those specialised arguments.

val run : ppf_dump:Format.formatter -> Flambda.set_of_closures -> Flambda.set_of_closures option